time axis is now a part of StreamLayout
